# microservice/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The gateway of forum",
        "title": "forum-gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "forum.muxixyz.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/collection/list/{user_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collection"
                ],
                "summary": "list收藏 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "last_id",
                        "name": "last_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.PostPartInfoResponse"
                        }
                    }
                }
            }
        },
...
        "/comment/{comment_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "获取评论 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "comment_id",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comment.Comment"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "删除评论 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "comment_id",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
...
        "/feedback/image": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback"
                ],
                "summary": "上传反馈截图 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "feedback_image",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/FeedbackUploadImageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/post": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "修改帖子信息 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "update_info_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/post.UpdateInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "创建帖子 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "create_post_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/post.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.IdResponse"
                        }
                    }
                }
            }
        },
...
        "/post/{post_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "获取帖子 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "post_id",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.GetPostResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "删除帖子 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "post_id",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "quality=1表示将帖子从精华板块移除，quality=0或不传表示直接删除帖子",
                        "name": "quality",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/report": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "处理举报 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "handle_report_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/report.HandleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "举报帖子 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "create_report_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/report.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "FeedbackUploadImageResponse": {
            "type": "object",
            "properties": {
                "file_token": {
                    "type": "string"
                }
            }
        },
...
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "comment.Comment": {
            "type": "object",
            "properties": {
                "be_replied_content": {
                    "type": "string"
                },
                "be_replied_user_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "creator_avatar": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "creator_name": {
                    "type": "string"
                },
                "father_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "like_num": {
                    "type": "integer"
                },
                "type_name": {
                    "description": "first-level -\u003e 一级评论; second-level -\u003e 其它级",
                    "type": "string"
                }
            }
        },
...
        "post.Comment": {
            "type": "object",
            "properties": {
                "be_replied_content": {
                    "type": "string"
                },
                "be_replied_id": {
                    "type": "integer"
                },
                "be_replied_user_id": {
                    "type": "integer"
                },
                "be_replied_user_name": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "creator_avatar": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "creator_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "like_num": {
                    "type": "integer"
                }
            }
        },
        "post.CreateRequest": {
            "type": "object",
            "required": [
                "category",
                "content",
                "content_type",
                "domain",
                "summary",
                "tags",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "compiled_content": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "description": "md or rtf",
                    "type": "string"
                },
                "domain": {
                    "description": "normal -\u003e 团队外; muxi -\u003e 团队内 (type_name暂时均填normal)",
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "post.GetPostResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "collection_num": {
                    "type": "integer"
                },
                "comment_num": {
                    "type": "integer"
                },
                "compiled_content": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "description": "md or rtf",
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "creator_avatar": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "creator_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_collection": {
                    "type": "boolean"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "like_num": {
                    "type": "integer"
                },
                "sub_posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/post.SubPost"
                    }
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "post.IdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "post.PostPartInfo": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "collection_num": {
                    "type": "integer"
                },
                "comment_num": {
                    "type": "integer"
                },
                "creator_avatar": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "creator_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "like_num": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "post.PostPartInfoResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/post.PostPartInfo"
                    }
                }
            }
        },
...
        "post.SubPost": {
            "type": "object",
            "properties": {
                "comment_num": {
                    "type": "integer"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/post.Comment"
                    }
                },
                "content": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "creator_avatar": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "creator_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "like_num": {
                    "type": "integer"
                }
            }
        },
...
        "post.UpdateInfoRequest": {
            "type": "object",
            "required": [
                "category",
                "content",
                "domain",
                "id",
                "tags",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "report.CreateRequest": {
            "type": "object",
            "required": [
                "cause",
                "id",
                "type_name"
            ],
            "properties": {
                "category": {
                    "description": "可选参数",
                    "type": "string"
                },
                "cause": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "id": {
                    "description": "post的id或者是comment的id",
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "type_name": {
                    "description": "post or comment",
                    "type": "string"
                }
            }
        },
        "report.HandleRequest": {
            "type": "object",
            "required": [
                "id",
                "result"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "result": {
                    "description": "invalid or valid",
                    "type": "string"
                }
            }
        },
...
    },
...
}