# microservice/gateway/handler/post/delete.go
package post

import (
	. "forum-gateway/handler"
	"forum-gateway/util"
	pb "forum-post/proto"
	"forum/log"
	"forum/model"
	"forum/pkg/constvar"
	"forum/pkg/errno"
	"strconv"

	"forum/client"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"
)

// Delete ... 删除帖子
// @Summary 删除帖子 api
// @Tags post
// @Accept application/json
// @Produce application/json
// @Param Authorization header string true "token 用户令牌"
// @Param post_id path int true "post_id"
// @Param quality query string false "quality=1表示将帖子从精华板块移除，quality=0或不传表示直接删除帖子"
// @Success 200 {object} Response
// @Router /post/{post_id} [delete]
func (a *Api) Delete(c *gin.Context) {
	log.Info("Post Delete function called.", zap.String("X-Request-Id", util.GetReqID(c)))

	userId := c.MustGet("userId").(uint32)

	id, err := strconv.Atoi(c.Param("post_id"))
	if err != nil {
		SendError(c, errno.ErrPathParam, nil, err.Error(), GetLine())
		return
	}

	ok, err := model.Enforce(userId, constvar.Post, id, constvar.Write)
	if err != nil {
		SendError(c, errno.ErrCasbin, nil, err.Error(), GetLine())
		return
	}

	if !ok {
		SendError(c, errno.ErrPermissionDenied, nil, "权限不足", GetLine())
		return
	}

	quality := c.DefaultQuery("quality", constvar.DeletePost)

	deleteReq := &pb.DeleteItemRequest{
		Id:       uint32(id),
		TypeName: constvar.Post,
		UserId:   userId,
	}

	if quality == constvar.RemoveQuality {
		ok, err := model.HasRole(userId, constvar.NormalAdminRole)
		if err != nil {
			SendError(c, errno.ErrCasbin, nil, err.Error(), GetLine())
			return
		}

		if !ok {
			SendError(c, errno.ErrPermissionDenied, nil, "权限不足", GetLine())
			return
		}

		deleteReq.TypeName = constvar.QualityPost
	}

	_, err = client.PostClient.DeleteItem(c.Request.Context(), deleteReq)
	if err != nil {
		SendError(c, err, nil, "", GetLine())
		return
	}

	SendResponse(c, nil, nil)
}